Moved all functions into a custom react hook, added wrapper functions in App.jsx to execute said react hook functions
