+++ src/App.jsx
@@ -2,6 +2,7 @@ import React, {useState, useRef} from "react";
 import styled from "styled-components";
 //import EditOutlinedIcon from '@mui/icons-material/EditOutlined';
 import CancelOutlinedIcon from '@mui/icons-material/CancelOutlined';
+import YoutubeDetailHook from "./Utils/YoutubeAPIFunctions";
  
 const Container = styled.div`
     width : 100vw;
@@ -198,127 +199,35 @@ function App() {
 
   const inputRef = useRef(null); //used to reference the api key input to get its value
 
-  /* Adds new row to list */
-  async function createNewStreamer(name, streamerChannelID, apiKey){
+  const [createNewStreamer, deleteRow, handleSubmit] = YoutubeDetailHook(); //custom hook functions
 
-    //get channel ID of streamer
-    const channelId = await getChannelID(streamerChannelID, apiKey)
-    .then((id) => {
-      return id;
-    }).catch(() => {
-      alert(`Could not fetch channel ID from Youtube API. Please make sure that you typed in a valid YouTube Channel ID.`);
-    });
-
-    //next, fetch JSON data about Youtube Channel's live stream status and save into variable
-    const streamStatus = await getData(channelId, apiKey).then((json) => {
-      return json.pageInfo.totalResults; //return json's pageInfo's totalResults attribute (either 0 or 1)
-    }).catch(() => {
-      alert(`Could not fetch data from Youtube API. Please try again.`)
-    });
-
-    //third, fetch channel's youtube profile picture 
-    const channelImageLink = await getChannelImage(channelId, apiKey).then((json) => {
-      //access json object to get the "default" thumbnail (88x88 px)
-      let imageURL = String(json.items[0].snippet.thumbnails.default.url); //get url as string
-      console.log(imageURL);
-      return imageURL; 
-    });
-
-    //fourth, create a streamer object depending on if the channel ID starts with '@' or not.
-    let newStreamer; 
-
-    //if channel ID starts with @
-    if(String(streamerChannelID).indexOf('@') === 0){
-      if(streamStatus === 1){ //if currently streaming
-        newStreamer = {id : Math.random(), name : name, streamStatus : "Currently Streaming", link: "https://youtube.com/".concat(streamerChannelID, "/live"), imgLink: channelImageLink};
-      }
-      else{ //if not currently streaming
-        newStreamer = {id : Math.random(), name : name, streamStatus : "Not Currently Streaming", link: "https://youtube.com/".concat(streamerChannelID, "/live"), imgLink: channelImageLink};
-      }
-    }
-
-    //else if channel ID doesn't start w/ '@'
-    else{
-      if(streamStatus === 1){
-        newStreamer = {id : Math.random(), name : name, streamStatus : "Currently Streaming", link: "https://youtube.com/channel/".concat(streamerChannelID, "/live"), imgLink: channelImageLink};
-      }    
-      else{
-        newStreamer = {id : Math.random(), name : name, streamStatus : "Not Currently Streaming", link: "https://youtube.com/channel/".concat(streamerChannelID, "/live"), imgLink: channelImageLink};
-      }
-    }
-
-    /*lastly, update the state of the component*/
-    //increment streamer count by 1
-    setStreamerCount((prevCount) => prevCount + 1);
-
-    //add to list of streamers
-    setStreamerList([...streamerList, newStreamer]);
+  //wrapper function for executing createNewStreamer function in our custom hook and setting our state values.
+  const handleCreateNewStreamer = async () => {
+    const streamerValues = await createNewStreamer(nameInput, streamerChannelInput, apiKey, streamerCount, streamerList); //returns an array that contains streamerCount and streamerList after running createNewStreamer() and the promise resolves.
     
+    //set our new streamer count 
+    setStreamerCount(streamerValues[0]); 
+
+    //set our new streamer list
+    setStreamerList(streamerValues[1]);
+      
     //reset nameInput
     setNameInput('');
-    
+            
     //reset streamerChannelInput
     setStreamerChannelInput('');
   };
 
-  /* Get YouTube channel's official Channel ID */
-  async function getChannelID(streamerChannelID, apiKey) {
+  //wrapper function for executing deleteRow function from custom hook and setting our state values.
+  const handleDeleteRow = (id) => {
+    //sets new streamer list and streamer count from array return from deleteRow function in custom hook
+    const deleteRowValues = deleteRow(id, streamerList, streamerCount);
 
-    //we are searching the youtube API w/ inputs of type="channel", fields="items%2Fsnippet%2FchannelId", q=${streamerChannelID}, key=${apiKey}, and maxResults="1" that returns a JSON object that contains the requested youtube channel ID
-    let response = await fetch(`https://www.googleapis.com/youtube/v3/search?part=snippet&type=channel&fields=items%2Fsnippet%2FchannelId&q=${streamerChannelID}&key=${apiKey}&maxResults=1`); 
-    
-    //convert json into text
-    const text = await response.text();
-    
-    //parse into js object
-    let jsonObject = JSON.parse(text); 
+    //set new streamer count
+    setStreamerCount(deleteRowValues[0]);
 
-    //return promise w/ channel ID as string
-    return jsonObject.items[0].snippet.channelId;
-  }
-
-  /* Get YouTube channel's Profile Image */
-  async function getChannelImage(channelId, apiKey){
-    //fetch from API for json data containing image link
-    let channelImage = await fetch(`https://www.googleapis.com/youtube/v3/channels?part=snippet&id=${channelId}&fields=items(id%2Csnippet%2Fthumbnails)&key=${apiKey}`); 
-
-    //convert response to json
-    const json = await channelImage.json(); 
-
-    //return json as promise
-    return json; 
-  }
-
-  /* Get YouTube channel's livestreaming status */
-  async function getData(channelID, apiKey){
-    try{
-      //search the youtube API for videos that are currently live (Livestreams) based on channel ID
-      const response = await fetch(`https://www.googleapis.com/youtube/v3/search?part=snippet&channelId=${channelID}&type=video&eventType=live&key=${apiKey}`);
-
-      //convert data into JS object
-      const json = await response.json() 
-
-      //return promise w/ js object
-      return json; 
-    }
-    catch{
-      console.log("Something went wrong!");
-    }
-  }
-
-
-  /* Delete Row */
-  const deleteRow = (id) => {
-    //filter out streamer list using array.filter(), adding each streamer to the new array that doesn't equal the id of the streamer row we want to delete
-    const newStreamerList = streamerList.filter((streamer) => streamer.id !== id);
-
-    setStreamerList(newStreamerList);
-
-    setStreamerCount((prevCount) => prevCount - 1);
-  };
-
-  const handleSubmit = (apiKey) => {
-    setAPIKey(apiKey)
+    //set new streamer list
+    setStreamerList(deleteRowValues[1]);
   };
 
   return (
@@ -327,7 +236,7 @@ function App() {
 
         <ApiKeyDiv style={apiKey.length > 0 ? {display : "none"} : {display : "flex"}}>
           <ApiKeyInput type="text" placeholder="Please enter your YouTube API key." ref={inputRef}/>
-          <ApiKeySubmitButton type="button" value="Submit" onClick={() => {handleSubmit(inputRef.current.value);}}/>
+          <ApiKeySubmitButton type="button" value="Submit" onClick={() => {setAPIKey(handleSubmit(inputRef));}}/>
           <InstructionsDiv>
             <InstructionsTitle>Instructions for creating your YouTube API Key (PLEASE READ!)</InstructionsTitle>
             <InstructionsSpan>
@@ -354,7 +263,7 @@ function App() {
         <AddStreamerButton style={apiKey.length === 0 ? {display : "none"} : {display : "initial"}} onClick={() => 
           {
             if(nameInput.length > 0 && streamerChannelInput.length > 0) 
-              {createNewStreamer(nameInput, streamerChannelInput, apiKey)}             
+              {handleCreateNewStreamer()}             
             else {
               alert("Please input the streamer's name and their YouTube channel ID.")
               return false;
@@ -377,7 +286,7 @@ function App() {
                 <StreamerName>{item.name}</StreamerName>
                 <StreamerStatus href={item.link} target="_blank">{item.streamStatus}</StreamerStatus>         
                 <ButtonDiv>
-                  <CancelButton onClick={() => deleteRow(item.id)}/>
+                  <CancelButton onClick={() => handleDeleteRow(item.id)}/>
                 </ButtonDiv>     
               </StreamerRow>)
           })
